% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_grid.R
\name{add_regular_polygon_grid}
\alias{add_regular_polygon_grid}
\title{Complete a data.frame with the quadkeys needed to have a regular grid.}
\usage{
add_regular_polygon_grid(data)
}
\arguments{
\item{data}{A data.frame with a quadkey column.}
}
\value{
An sf POLYGON data.frame with all the quadkeys within
the bounding box of the provided quadkeys.
The columns quadkey and geometry are returned for all the grid,
and the values of the introduced quadkeys will be NA
for the rest pf the variables.
}
\description{
This function estimates the bounding box of the quadkeys given in the
quadkey column and adds rows to complete the quadkeys and the geometry
needed to create a regular grid.
All other columns for the introduced quadkeys will be filled with NAs.
}
\examples{
# read the file with the data
path <- paste0(system.file("extdata", package = 'quadkeyr'),
                             "/cityA_2020_04_15_0000.csv")
data <- read.csv(path)
data <- format_fb_data(data)

get_regular_polygon_grid(data = data)
}
