% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_to_polygon.R
\name{complete_grid_for_polygons}
\alias{complete_grid_for_polygons}
\title{Prepares the grid of QuadKeys for the conversion to square polygons}
\usage{
complete_grid_for_polygons(data)
}
\arguments{
\item{data}{A spatial dataset (sf) with the columns tileX, tileY and quadkey.}
}
\value{
A spatial dataset (sf) with the columns tileX, tileY and quadkey.
}
\description{
The QuadKey's points of the grid represent the upper-left corner
of the QuadKey.
This function creates an extra tile row and column needed to create
the square polygons for the QuadKeys that are in the the borders of the area
without adding explicitly new QuadKeys to the grid.
This function is called internally by grid_to_polygon.
}
\examples{

grid  <-  create_qk_grid(xmin = -59,
                      xmax = -57,
                      ymin = -35,
                      ymax = -34,
                      level = 11)

grid_coords <- get_qk_coord(data = grid$data)


# You can use grid_to_polygon

extragrid <- complete_grid_for_polygons(grid_coords)


}
\seealso{
\code{\link{grid_to_polygon}}
}
