% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_coords.R
\name{extract_tile_coord}
\alias{extract_tile_coord}
\title{Extract lat/long coordinates from the tile XY coordinates.}
\usage{
extract_tile_coord(data, level)
}
\arguments{
\item{data}{A dataframe with columns named tileX and tileY}

\item{level}{Level of detail, from 1 (lowest detail) to 23 (highest detail).}
}
\value{
A spatial dataframe (sf) containing the tiles XY coordinates
(tileX, tileY)  and a column for POINT geometry.
}
\description{
Reads the tile XY coordinates and extracts the
lat/long coordinates of the upper-left corner of the QuadKey.
}
\examples{


grid <- create_qk_grid(xmin = -59,
                       xmax = -40,
                       ymin = -38,
                       ymax = -20,
                       level = 6)

extract_tile_coord(data = grid$data,
                   level = 6)

}
\seealso{
\code{\link{tileXY_to_pixelXY}}

\code{\link{pixelXY_to_latlong}}
}
