% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_format_files.R
\name{format_fb_data}
\alias{format_fb_data}
\title{Format the data}
\usage{
format_fb_data(data)
}
\arguments{
\item{data}{A dataframe with a quadkey, date_time, country columns and
other numeric variables}
}
\value{
A dataframe.
}
\description{
This function removes unnecessary characters such as \verb{\\\\N}
and ensures that the format of the date and QuadKeys is correct.
}
\examples{

data(result_read_fb_mobility_data)
format_fb_data(data = result_read_fb_mobility_data)
}
\seealso{
\code{\link{read_fb_mobility_files}}
}
