% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_grid.R
\name{get_regular_polygon_grid}
\alias{get_regular_polygon_grid}
\title{Create a data.frame with all the quadkeys needed to
generate a regular grid.}
\usage{
get_regular_polygon_grid(data)
}
\arguments{
\item{data}{A data.frame with a quadkey column.}
}
\value{
An sf POLYGON data.frame with all the quadkeys within
the bounding box of the provided quadkeys.
Only the columns quadkey and geometry are returned.
}
\description{
This function estimates the bounding box of the quadkeys given in the
quadkey column and adds rows to complete the quadkeys and the geometry
needed to create a regular grid.
}
\examples{

# daya file
path <- paste0(system.file("extdata", package = 'quadkeyr'),
                             "/cityA_2020_04_15_0000.csv")
data <- read.csv(path)
data <- format_fb_data(data)

get_regular_polygon_grid(data = data)
}
