% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coords.R
\name{get_tile_coord}
\alias{get_tile_coord}
\title{Get lat/long coordinates from the tile XY coordinates.}
\usage{
get_tile_coord(data, zoom)
}
\arguments{
\item{data}{A dataframe with columns named \code{tileX} and \code{tileY}}

\item{zoom}{Zoom or Level of detail,
from 1 (lowest detail) to 23 (highest detail).}
}
\value{
A \code{sf} POINT data.frame containing the tiles XY coordinates
(\code{tileX}, \code{tileY}), a \code{quadkey} and a \code{geometry} column.
}
\description{
Reads the tile XY coordinates and extracts the
lat/long coordinates of the upper-left corner of the QuadKey.
}
\examples{


grid <- create_qk_grid(
  xmin = -59,
  xmax = -40,
  ymin = -38,
  ymax = -20,
  zoom = 6
)

# tileX and tileY columns in grid$data converted to geographic coordinates
get_tile_coord(
  data = grid$data,
  zoom = 6
)

}
\seealso{
\code{\link{tileXY_to_pixelXY}}

\code{\link{pixelXY_to_latlong}}
}
\keyword{internal}
