% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong_to_qk_functions.R
\name{latlong_to_pixelXY}
\alias{latlong_to_pixelXY}
\title{Converts lat/long coordinates to pixel XY coordinates}
\usage{
latlong_to_pixelXY(latitude, longitude, level)
}
\arguments{
\item{latitude}{Latitude of the point, in degrees.}

\item{longitude}{Longitude of the point, in degrees.}

\item{level}{Level of detail, from 1 (lowest detail) to 23 (highest detail).}
}
\value{
A list returning pixel X and pixel Y coordinates.
}
\description{
Converts a point from latitude/longitude WGS-84 coordinates (in degrees) into pixel XY coordinates at a specified level of detail.
For further information, refer to the Microsoft Bing Maps Tile System documentation.
}
\examples{

latlong_to_pixelXY(latitude = -35,
                 longitude = -50,
                 level = 6)

}
