% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{mapsize}
\alias{mapsize}
\title{Map size in pixels}
\usage{
mapsize(level)
}
\arguments{
\item{level}{Level of detail, from 1 (lowest detail) to 23 (highest detail).}
}
\value{
The map width and height in pixels.
}
\description{
Determines the map width and height (in pixels) at a specified
level of detail.
For further information, refer to the Microsoft Bing Maps Tile
System documentation.
}
\examples{

mapsize(level = 6)

}
\seealso{
\href{https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system}{
Microsoft Bing Maps Tile System documentation}
}
