% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_to_raster.R
\name{polygon_to_raster}
\alias{polygon_to_raster}
\title{Create and save raster images for different dates and times}
\usage{
polygon_to_raster(
  data,
  nx,
  ny,
  template,
  variable = "percent_change",
  filename,
  path
)
}
\arguments{
\item{data}{A spatial dataframe (sf) with the variable we want to represent
in the raster.}

\item{nx}{Integer; number of cells in x direction.}

\item{ny}{Integer; number of cells in y direction.}

\item{template}{A spatial dataset (sf) with the polygon grid used as
template}

\item{variable}{The column name of the variable to plot.}

\item{filename}{Select a name for the file. The date and time will
be included automatically in the name.}

\item{path}{Path where the files should be stored.}
}
\value{
as many .tif files as dates and times in the dataset
}
\description{
Creates one raster by each date and time reported and
saves it as a .tif.
}
\examples{
\dontrun{

files <- read_fb_mobility_files(path_to_csvs = "../geocovid/data/rasters/city/",
 colnames = c("lat", "lon", 
            "quadkey", "date_time", 
            "n_crisis", "percent_change"),
. coltypes = list(
 lat = 'd',
 lon = 'd',
 quadkey = 'i',
 date_time = 'T',
 n_crisis = 'c',
 percent_change = 'c')) 

# Extract unique QuadKey values
quadkeys <-  unique(files$quadkey)

# Convert QuadKeys to geographic coordinates
qtll <- quadkey_to_latlong(quadkeys = quadkeys)

# Complete the grid (if necessary)
regular_grid <- regular_qk_grid(qtll)

# Create polygons
polygrid  <-  grid_to_polygon(grid_coords)

# Keep only the QuadKeys reported
polyvar <- files |>
           dplyr::inner_join(polygrid, by = 'quadkey' )

# Generate the raster files                       
polygon_to_raster(data = polyvar,
                  nx = grid$num_cols + 1,
                  ny = grid$num_rows + 1,
                  template = polyvar,
                  variable = 'percent_change',
                  filename = 'cityA',
                  path = "data/")
}

}
\seealso{
\code{\link{st_as_stars}}, \code{\link{st_rasterize}}

\code{\link{missing_combinations}}
}
