% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qkmap_app.R
\name{qkmap_app}
\alias{qkmap_app}
\title{Launch the Shiny App}
\usage{
qkmap_app()
}
\value{
This function does not return any value. It launches the Shiny app.
}
\description{
This function launches the Shiny app.
}
\examples{
\dontrun{
qkmap_app()
}
}
