% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qk_to_polygon_functions.R
\name{quadkey_df_to_polygon}
\alias{quadkey_df_to_polygon}
\title{Convert data.frame with quadkey column to a sf POLYGON data.frame}
\usage{
quadkey_df_to_polygon(data)
}
\arguments{
\item{data}{A data.frame with a quadkey column}
}
\value{
The same original data.frame with a sf POLYGON data.frame with a
geometry column.
}
\description{
Convert data.frame with quadkey column to a sf POLYGON data.frame
}
\examples{

# get data file
path <- paste0(system.file("extdata", package = 'quadkeyr'),
                                    "/cityA_2020_04_15_0000.csv")
data <- read.csv(path)
data <- format_fb_data(data)

quadkey_df_to_polygon(data = data)
}
