% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadkeyr-package.R
\docType{package}
\name{quadkeyr-package}
\alias{quadkeyr}
\alias{quadkeyr-package}
\title{quadkeyr: Tools for converting QuadKey-identified datasets (Microsoft's Bing Maps Tile System) into raster images}
\description{
'Quadkeyr' functions generate raster images based on QuadKey-identified data, facilitating efficient integration of Tile Maps data into R workflows. In particular, 'Quadkeyr' provides support to process and analyze Facebook mobility datasets within the R environment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fernandez-lab-wsu.github.io/quadkeyr/}
  \item \url{https://github.com/Fernandez-Lab-WSU/quadkeyr}
  \item Report bugs at \url{https://github.com/Fernandez-Lab-WSU/quadkeyr/issues}
}

}
\author{
\strong{Maintainer}: Florencia D'Andrea \email{florencia.dandrea@gmail.com} (\href{https://orcid.org/0000-0002-0041-097X}{ORCID})

Authors:
\itemize{
  \item Pilar Fernandez \email{pilar.fernandez@wsu.edu} (\href{https://orcid.org/0000-0001-8645-2267}{ORCID})
}

Other contributors:
\itemize{
  \item Maria Paula Caldas (Maria Paula Caldas reviewed the package (v. 0.0.0.9000) for rOpenSci, see https://github.com/ropensci/software-review/issues/619) [reviewer]
  \item Vincent Van Hess (Vincent Van Hess reviewed the package (v. 0.0.0.9000) for rOpenSci, see https://github.com/ropensci/software-review/issues/619) [reviewer]
  \item CDC's Center for Forecasting and Outbreak Analytics (This project was made possible by cooperative agreement CDC-RFA-FT-23-0069 (grant # NU38FT000009-01-00) from the CDC's Center for Forecasting and Outbreak Analytics. Its contents are solely the responsibility of the authors and do not necessarily represent the official views of the Centers for Disease Control and Prevention.) [funder]
  \item MIDAS-NIH COVID-19 urgent grant program [funder]
  \item Washington State University [copyright holder]
}

}
\keyword{internal}
