% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_format_files.R
\name{read_all_files}
\alias{read_all_files}
\title{Read all the .csv files in a folder and format the data.}
\usage{
read_all_files(path_to_csvs = path_to_csvs)
}
\arguments{
\item{path_to_csvs}{Path to the folder where the .csv files are stored}
}
\value{
A dataframe with the information of all the files.
}
\description{
This function reads all the .csv files in a particular folder.
These files consistently contain identical columns, with variations
only in location, day, and time. As a result, we can uniformly apply
specific formatting to columns across these files."
}
\examples{


# read_all_files(path_to_csvs = 'data/')

}
\seealso{
\code{\link{format_data}}
}
