% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regular_grid.R
\name{regular_qk_grid}
\alias{regular_qk_grid}
\title{Converts a irregular QuadKey grid into a regular grid}
\usage{
regular_qk_grid(data)
}
\arguments{
\item{data}{A spatial dataset of class sf with POINT geometry}
}
\value{
A spatial dataset of class sf with POINT geometry
}
\description{
This function completes the grid of QuadKeys using the bounding
box of the data provided.
}
\examples{

quadkeys <- c('213', '210', '211')

qtll <- quadkey_to_latlong(quadkeys = quadkeys)

regular_qk_grid(qtll)

}
\seealso{
\code{\link{create_qk_grid}}

\code{\link{quadkey_to_latlong}}
}
