% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{result_read_fb_mobility_data}
\alias{result_read_fb_mobility_data}
\title{result_read_fb_mobility_data: Fake dataset}
\format{
\subsection{\code{result_read_fb_mobility_data}}{

A data frame with 134,492 rows and 9 columns:
\describe{
\item{lat}{Latitude of QuadKey centroid}
\item{lon}{Longitude of QuadKey centroid}
\item{quadkey}{QuadKey number as character}
\item{country}{Country name}
\item{date_time}{Date in format \%Y-\%m-\%d \%H\%M}
\item{n_crisis}{Variable}
\item{percent_change}{Variable}
\item{day}{Day in format \%Y-\%m-\%d}
\item{hour}{Hour of the day as a number between 1 and 24}
}
}
}
\usage{
result_read_fb_mobility_data
}
\description{
A data frame similar to a the potential output of read_all_files().
}
\keyword{datasets}
