% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong_to_qk_functions.R
\name{tileXY_to_quadkey}
\alias{tileXY_to_quadkey}
\title{Converts tile XY coordinates into a quadkey.}
\usage{
tileXY_to_quadkey(tileX, tileY, level)
}
\arguments{
\item{tileX}{Tile X coordinate.}

\item{tileY}{Tile Y coordinate.}

\item{level}{Level of detail, from 1 (lowest detail) to 23 (highest detail).}
}
\value{
The quadkey number as a string.
}
\description{
Converts tile XY coordinates into a QuadKey at a specified
level of detail.
For further information, refer to the Microsoft Bing Maps Tile System
documentation.
}
\examples{

tileXY_to_quadkey(tileX = 23,
                  tileY = 38,
                  level = 6)

}
\seealso{
\href{https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system}{
Microsoft Bing Maps Tile System documentation}
}
