% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_weekly_lag.R
\name{apply_weekly_lag}
\alias{apply_weekly_lag}
\title{Apply a 7 day lag to the variable n_crisis}
\usage{
apply_weekly_lag(data)
}
\arguments{
\item{data}{A dataset}
}
\value{
A dataframe with the extra colums n_crisis_lag_7 and n_crisis_percent_7
}
\description{
Applying a week lag to the data will create rasters showing how
people was moving a week ago the date of interest. This function works only
for quadkeys without NAs.
}
\examples{

data <- result_read_all_files[1:500,]
apply_weekly_lag(data)

}
