% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_weekly_lag.R
\name{apply_weekly_lag}
\alias{apply_weekly_lag}
\title{Apply a 7 day lag to the variable n_crisis}
\usage{
apply_weekly_lag(data)
}
\arguments{
\item{data}{A dataset}
}
\value{
A dataframe with the extra columns n_crisis_lag_7 and
percent_change_7.

n_crisis_lag_7, is the same variable defined as n_crisis in the Facebook
dataset with a 7 day lag applied.

percent_change_7 is the difference between the n_crisis value between weeks
expressed as percentage.
}
\description{
Applying a week lag to the data will create raster images
showing the mobility a week before the date of interest.
This function works only for quadkeys without NAs.
}
\examples{

data <- result_read_fb_mobility_data[1:500, ]
apply_weekly_lag(data)

}
