% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_to_polygon.R
\name{complete_grid_for_polygons}
\alias{complete_grid_for_polygons}
\title{Prepare the grid of QuadKeys for the conversion to square polygons}
\usage{
complete_grid_for_polygons(data)
}
\arguments{
\item{data}{A \code{sf} POINT data.frame with the \code{quadkey},
\code{tileX}, \code{tileY} and \code{geometry} columns}
}
\value{
A data.frame with the \code{tileX}, \code{tileY} and \code{quadkey} columns.
QuadKeys appears as NAs as there are only kept to extend the grid.
Their value is not necessary.
}
\description{
The QuadKey's points of the grid represent
the upper-left corner of the QuadKey.
This function creates an extra tile row and column needed to create
the square polygons for the QuadKeys
that are in the the borders of the area
without adding explicitly new QuadKeys to the grid.
This function is called internally by \code{\link[=grid_to_polygon]{grid_to_polygon()}}.
}
\examples{

grid <- create_qk_grid(
  xmin = -59,
  xmax = -57,
  ymin = -35,
  ymax = -34,
  zoom = 11
)

grid_coords <- get_qk_coord(data = grid$data)

extragrid <- quadkeyr::complete_grid_for_polygons(grid_coords)
extragrid
}
\seealso{
\code{\link{grid_to_polygon}}
}
\keyword{internal}
