% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_provided}
\alias{data_provided}
\title{data_provided: fake dataset}
\format{
\subsection{\code{data_provided}}{

A data frame with 360 rows and 2 columns:
\describe{
\item{quadkey}{QuadKey as character}
\item{variable}{Numeric variable}
}
}
}
\usage{
data_provided
}
\description{
A data frame simulating a variable asociated with QuadKey numbers.
}
\keyword{datasets}
