% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_format_files.R
\name{missing_combinations}
\alias{missing_combinations}
\title{Detect dates and hours missing in filenames}
\usage{
missing_combinations(data)
}
\arguments{
\item{data}{A data.frame with a \code{day} and hour column.}
}
\value{
A data.frame with the missing days and hours, if any.
}
\description{
Facebook mobility data is reported daily at 3 different hours
(0, 8, 16).
This function reads the data extracted from the current files and detects
if any file for a particular day or hour is missing.
}
\examples{

# Sample dataset
data <- data.frame(
  country = c("US", "MX", "MX"), 
  day = c("2023-01-01", "2023-01-03", "2023-01-05"),
  hour = c(0, 8, 16)
)

missing_combinations(data)
}
