% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_format_files.R
\name{missing_combinations}
\alias{missing_combinations}
\title{Detects dates and times missing}
\usage{
missing_combinations(data)
}
\arguments{
\item{data}{A dataframe with a day and time column.}
}
\value{
A dataframe with the missing days and times, if any.
}
\description{
Facebook mobility data is reported daily at 3 different times
(0, 8, 16).
This function reads the data extracted from the current files and detects
if any day or time is missing.
}
\examples{

# Sample dataset
data <- data.frame(
 day = c("2023-01-01", "2023-01-03", "2023-01-05"),
 hour = c(0, 8, 16)
)

missing_combinations(data)

}
