% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latlong_to_qk_functions.R
\name{pixelXY_to_tileXY}
\alias{pixelXY_to_tileXY}
\title{Converts pixel XY coordinates into tile XY coordinates}
\usage{
pixelXY_to_tileXY(pixelX, pixelY)
}
\arguments{
\item{pixelX}{Pixel X coordinate.}

\item{pixelY}{Pixel Y coordinate.}
}
\value{
A list returning the tile X and tile Y coordinates.
}
\description{
Converts pixel XY coordinates into tile XY coordinates of the
tile containing the specified pixel.
For further information, refer to the Microsoft Bing Maps
Tile System documentation.
}
\examples{

pixelXY_to_tileXY(
  pixelX = 5916,
  pixelY = 9894
)

}
\seealso{
\href{
https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system
}{Microsoft Bing Maps Tile System documentation}
}
