% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_to_raster.R
\name{polygon_to_raster}
\alias{polygon_to_raster}
\title{Create and save raster images for different dates and times}
\usage{
polygon_to_raster(
  data,
  nx,
  ny,
  template,
  variable = "percent_change",
  filename,
  path
)
}
\arguments{
\item{data}{A spatial dataframe (sf) with the variable we want to represent
in the raster.}

\item{nx}{Integer; number of cells in x direction.}

\item{ny}{Integer; number of cells in y direction.}

\item{template}{A spatial dataset (sf) with the polygon grid used as
template}

\item{variable}{The column name of the variable to plot.}

\item{filename}{Select a name for the file. The date and time will
be included automatically in the name.}

\item{path}{Path where the files should be stored.}
}
\value{
as many .tif files as dates and times in the dataset
}
\description{
Creates one raster by each date and time reported and
saves it as a .tif.
}
\examples{

# TBD

}
\seealso{
\code{\link{st_as_stars}}, \code{\link{st_rasterize}}

\code{\link{missing_combinations}}
}
