% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qk_to_latlong_functions.R
\name{quadkey_to_latlong}
\alias{quadkey_to_latlong}
\title{Convert a string of Quadkey numbers to lat/long coordinates}
\usage{
quadkey_to_latlong(quadkeys)
}
\arguments{
\item{quadkeys}{A vector with unique QuadKey numbers.}
}
\value{
A spatial data frame of class sf with a quadkey column
and POINT geometry.
The coordinates represent the upper-left corner of the QuadKey.
}
\description{
This function converts Quadkeys to
latitude/longitude WGS-84 coordinates (in degrees).
For further information, refer to the Microsoft Bing Maps Tile System
documentation.
}
\examples{

quadkey_to_latlong(c("213", "212", "210"))

}
\seealso{
\code{\link{quadkey_to_tileXY}}

\code{\link{tileXY_to_pixelXY}}

\code{\link{pixelXY_to_latlong}}

\href{https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system}{
Microsoft Bing Maps Tile System documentation}
}
