% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_to_polygon.R
\name{quadkey_to_polygon}
\alias{quadkey_to_polygon}
\title{Convert a QuadKey into a square polygon}
\usage{
quadkey_to_polygon(quadkey)
}
\arguments{
\item{quadkey}{The QuadKey as a string}
}
\value{
A spatial dataset (sf) with a quadkey and POLYGON geometry column.
}
\description{
This functions creates a sf class polygon from a QuadKey.
}
\examples{

quadkey_to_polygon('213')

}
