% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qk_to_latlong_functions.R
\name{quadkey_to_tileXY}
\alias{quadkey_to_tileXY}
\title{Converts a QuadKey into tile XY coordinates.}
\usage{
quadkey_to_tileXY(qk)
}
\arguments{
\item{qk}{The QuadKey number as a string.}
}
\value{
A list returning the tile X, tile Y coordinates and
the level of detail.
}
\description{
For further information, refer to the
Microsoft Bing Maps Tile System documentation.
}
\examples{

quadkey_to_tileXY(qk = 213)
}
\seealso{
\href{https://learn.microsoft.com/en-us/bingmaps/articles/bing-maps-tile-system}{
Microsoft Bing Maps Tile System documentation}
}
