% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadkeyr-package.R
\docType{package}
\name{quadkeyr-package}
\alias{quadkeyr}
\alias{quadkeyr-package}
\title{quadkeyr: Generate Raster Images From QuadKey-Identified Datasets}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of functions of increasing complexity allows users to (1) convert QuadKey-identified datasets, based on Microsoft's Bing Maps Tile System, into Simple Features data frames, (2) transform Simple Features data frames into rasters, and (3) process multiple Meta (Facebook) QuadKey-identified human mobility files directly into raster files. For more details, see D’Andrea et al. (2024) \doi{10.21105/joss.06500}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/quadkeyr/}
  \item \url{https://github.com/ropensci/quadkeyr}
  \item Report bugs at \url{https://github.com/ropensci/quadkeyr/issues}
}

}
\author{
\strong{Maintainer}: Florencia D'Andrea \email{florencia.dandrea@gmail.com} (\href{https://orcid.org/0000-0002-0041-097X}{ORCID})

Authors:
\itemize{
  \item Pilar Fernandez \email{pilar.fernandez@wsu.edu} (\href{https://orcid.org/0000-0001-8645-2267}{ORCID})
}

Other contributors:
\itemize{
  \item Maria Paula Caldas (\href{https://orcid.org/0000-0002-1938-6471}{ORCID}) (Maria Paula Caldas reviewed the package (v. 0.0.0.9000) for rOpenSci, see https://github.com/ropensci/software-review/issues/619) [reviewer]
  \item Vincent van Hees (\href{https://orcid.org/0000-0003-0182-9008}{ORCID}) (Vincent van Hees reviewed the package (v. 0.0.0.9000) for rOpenSci, see https://github.com/ropensci/software-review/issues/619) [reviewer]
  \item CDC's Center for Forecasting and Outbreak Analytics (This project was made possible by cooperative agreement CDC-RFA-FT-23-0069 (grant # NU38FT000009-01-00) from the CDC's Center for Forecasting and Outbreak Analytics. Its contents are solely the responsibility of the authors and do not necessarily represent the official views of the Centers for Disease Control and Prevention.) [funder]
  \item MIDAS-NIH COVID-19 urgent grant program [funder]
  \item Paul G. Allen School for Global Health, Washington State University [copyright holder]
}

}
\keyword{internal}
