% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadkeyr-package.R
\docType{package}
\name{quadkeyr-package}
\alias{quadkeyr}
\alias{quadkeyr-package}
\title{quadkeyr: Tools for converting QuadKey-identified datasets (Microsoft's Bing Maps Tile System) into raster images}
\description{
'Quadkeyr' functions generate raster images based on QuadKey-identified data, facilitating efficient integration of Tile Maps data into R workflows. In particular, 'Quadkeyr' provides support to process and analyze Facebook mobility datasets within the R environment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fernandez-lab-wsu.github.io/quadkeyr/}
  \item \url{https://github.com/Fernandez-Lab-WSU/quadkeyr}
  \item Report bugs at \url{https://github.com/Fernandez-Lab-WSU/quadkeyr/issues}
}

}
\author{
\strong{Maintainer}: Florencia D'Andrea \email{florencia.dandrea@gmail.com} (\href{https://orcid.org/0000-0002-0041-097X}{ORCID})

}
\keyword{internal}
