% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qk_to_latlong_functions.R
\name{tileXY_to_pixelXY}
\alias{tileXY_to_pixelXY}
\title{Convert tile XY coordinates into pixel XY coordinates}
\usage{
tileXY_to_pixelXY(tileX, tileY)
}
\arguments{
\item{tileX}{Tile X coordinate.}

\item{tileY}{Tile Y coordinate.}
}
\value{
A list returning the pixel X and pixel Y coordinates.
}
\description{
Converts tile XY coordinates into pixel XY coordinates of the upper-left pixel of the specified tile.
For further information, refer to the Microsoft Bing Maps Tile System documentation.
}
\examples{

tileXY_to_pixelXY(tileX = 3,
                  tileY = 5)

}
